Move image post and retirevel routes to controllers; add comments and remove unnecessary logging
