+++ server/server.js
@@ -6,20 +6,9 @@ require("dotenv").config();
 const fileUpload = require("express-fileupload");
 app.use(fileUpload());
 
-const s3Client = require("./config/s3setup");
-// Pass s3client to s3 routes
-require("./routes/POST-upload.js")(app, s3Client);
-require("./routes/GET-image.js")(app, s3Client);
-
 // app.use(cors());
 app.use(express.json()); // Parse JSON request bodies with an increased limit
 app.use(express.urlencoded({ extended: true })); // Parse URL-encoded data with an increased limit
-
-// app.use(session({
-//   secret: 'redissecretcloud',//newly-added, copied from .env, highl unsure
-//   resave: false,
-//   saveUninitialized: true
-// }));
 
 // TODO: Configure database connections, authentication, and other application-level settings
